Fixed a ton of small bugs. Getting there... this compiles, but doesn't link.

+++ FluidNC/src/Machine/MachineConfig.h
@@ -84,7 +84,7 @@ namespace Machine {
         bool  _verboseErrors     = true;
         bool  _reportInches      = false;
 
-        size_t _planner_blocks = 16;
+        uint32_t _planner_blocks = 16;
 
         // Enables a special set of M-code commands that enables and disables the parking motion.
         // These are controlled by `M56`, `M56 P1`, or `M56 Px` to enable and `M56 P0` to disable.
